# docs/api.yaml
openapi: 3.0.3
info:
  title: EventSync API
  description: API for event management and real-time participant engagement
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Development server
paths:
...
  /sessions:
    get:
      summary: List all sessions
      tags:
        - Sessions
      parameters:
        - name: room
          in: query
          schema:
            type: string
          description: Filter by room name
        - name: speaker
          in: query
          schema:
            type: string
          description: Filter by speaker name (matches first or last name)
        - name: live
          in: query
          schema:
            type: boolean
          description: Filter sessions currently live
        - name: event
          in: query
          schema:
            type: string
          description: Filter by event title
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number (starts at 1)
        - name: size
          in: query
          schema:
            type: integer
            default: 20
          description: Number of items per page
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionList'
        '422':
          $ref: '#/components/responses/Unprocessable'
    post:
      summary: Create a new session
      tags:
        - Sessions
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionInput'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
...
  /sessions/{id}/questions:
    get:
      summary: List questions of a session
      tags:
        - Questions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the session
        - name: sort
          in: query
          schema:
            type: string
            enum: [upvotes, createdAt]
            default: upvotes
          description: Sort order (upvotes descending, or createdAt ascending)
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number (starts at 1)
        - name: size
          in: query
          schema:
            type: integer
            default: 20
          description: Number of items per page
        - name: title
          in: query
          schema:
            type: string
          description: Filter questions by title (partial match, case-insensitive)
      responses:
        '200':
          description: List of questions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionList'
        '422':
          $ref: '#/components/responses/Unprocessable'
    post:
      summary: Ask a question on a session
      tags:
        - Questions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: ID of the session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionInput'
      responses:
        '201':
          description: Question created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Question'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required:
        - status
        - error
        - message
      properties:
        status:
          type: integer
          description: HTTP status code
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: Field affected by the error
              message:
                type: string
                description: Error description for this field
...
    EventRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
...
    Session:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        room:
          $ref: '#/components/schemas/Room'
        capacity:
          type: integer
        event:
          $ref: '#/components/schemas/EventRef'
        speakers:
          type: array
          items:
            $ref: '#/components/schemas/SpeakerRef'
        isLive:
          type: boolean
          description: True if the session is live
...
    SessionList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Session'
        meta:
          type: object
          properties:
            total:
              type: integer
            page:
              type: integer
            size:
              type: integer
    SessionInput:
      type: object
      required:
        - title
        - description
        - startDate
        - endDate
        - roomId
        - capacity
        - eventId
      properties:
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        roomId:
          type: string
          format: uuid
        capacity:
          type: integer
        eventId:
          type: string
          format: uuid
...
    Room:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
...
    SpeakerRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        profilePicture:
          type: string
          format: uri
        bio:
          type: string
...
    ParticipantRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
...
    Question:
      type: object
      required:
        - id
        - title
        - content
        - participant
        - upvotes
        - isAnonymous
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        participant:
          $ref: '#/components/schemas/ParticipantRef'
        upvotes:
          type: integer
        isAnonymous:
          type: boolean
        createdAt:
          type: string
          format: date-time
    QuestionList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Question'
        meta:
          type: object
          properties:
            total:
              type: integer
            page:
              type: integer
            size:
              type: integer
...
    QuestionInput:
      type: object
      required:
        - title
        - content
      properties:
        title:
          type: string
        content:
          type: string
        isAnonymous:
          type: boolean
          default: false

    # ────────────────── AI Conversation ──────────────────
...
  responses:
    BadRequest:
      description: Malformed request or invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            status: 400
            error: Bad Request
            message: The request body contains invalid data or a parameter is malformed
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            status: 401
            error: Unauthorized
            message: Authentication is required or the provided credentials are invalid
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            status: 404
            error: Not Found
            message: The requested resource was not found
    Conflict:
      description: Resource conflict (e.g. duplicate, state conflict)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            status: 409
            error: Conflict
            message: The request could not be completed due to a conflict with the current state of the resource
    Unprocessable:
      description: Semantically invalid data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            status: 422
            error: Unprocessable Entity
            message: At least one required field is missing or invalid
...